Added dummy getShowList request handler
For now this is just loading client/dev/data.js and parsing it
into a JSON da ta structure. Then re-serializing it back out in
the response.

In working on this, it was discovered that the JSON data we made
was malforme d, the keys of objects must be wrapped in quotes in
the serialized JSON.