Added ability to serialize the Programs data into JSON
Decided to switch JSON libraries from json to aeson and use the
DeriveGeneric behavior of GHC. This necessitated a large redesign
of the program and episode data structures as well.

Also, this required a change to the JSON object keys in our
Coffeescript. This happened because of Haskell's record-style
datatype limitation where field names must be unique.