Added conf file support and filtering our non-video files
Added a new module, Vidproj.Conf, to support loading of key/value style configuration. Moved the conf for http port and top level video dirs out of the source code.

This change also adds support for filtering the video files based on a list of extensions. This is also a conf file option. A dummy test file was added for this as well.